refactor(lock): add useLock Vue composable under @snapshot-labs/lock/vue
Move the connector-instantiation and EIP-6963 subscription logic out of
apps/ui's useConnectors composable and into a new useLock composable
exported from @snapshot-labs/lock/vue. The UI now just defines the
CONNECTOR_DETAILS config and delegates construction to useLock.

Also rename the package's barrel to export Connector as the type (was the
class) — class is now exported as ConnectorClass — and rewrite the README
with usage examples.
