+++ packages/lock/src/index.ts
@@ -1,5 +1,5 @@
 export { default as connectors } from './connectors';
-export { default as Connector } from './connectors/connector';
+export { default as ConnectorClass } from './connectors/connector';
 export { default as Eip6963 } from './eip6963';
 export { GuestProvider, formatAddress } from './connectors/guest';
-export type { Connector as ConnectorInstance, ConnectorType } from './types';
+export type { Connector, ConnectorType } from './types';
