+++ packages/lock/src/vue/index.ts
@@ -1,0 +1,2 @@
+export { useLock } from './useLock';
+export type { ConnectorDetail } from './useLock';
